backend/pyproject.toml
[tool.poetry]
name = "jobt-ai-career-coach"
version = "1.0.0"
description = "AI-powered interview practice platform for African professionals"
authors = ["Micheal Olorundare <contact@jobt.ai>"]
license = "MIT"
readme = "README.md"
homepage = "https://jobt.ai"
repository = "https://github.com/yourusername/jobt-ai"
documentation = "https://docs.jobt.ai"
keywords = ["interview", "ai", "career", "coaching", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.12.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
openai = "^1.3.5"
httpx = "^0.25.2"
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.12.0"
isort = "^5.13.2"
mypy = "^1.7.1"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"

[tool.poetry.scripts]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8010"
start = "uvicorn app.main:app --host 0.0.0.0 --port 8010"
migrate = "alembic upgrade head"
seed = "python scripts/seed_categories.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ==================== Black Configuration ====================
[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | venv
  | alembic
  | build
  | dist
)/
'''

# ==================== isort Configuration ====================
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/alembic/*", "*/venv/*", "*/.venv/*"]

# ==================== pytest Configuration ====================
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# ==================== Coverage Configuration ====================
[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/alembic/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# ==================== mypy Configuration ====================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "alembic.*",
    "sqlalchemy.*",
    "asyncpg.*",
]
ignore_missing_imports = true

# ==================== Pydantic Plugin ====================
[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

# ==================== Ruff Configuration (Alternative to flake8) ====================
[tool.ruff]
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".venv",
    "venv",
    "alembic",
    "build",
    "dist",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

# ==================== Bandit Configuration (Security) ====================
[tool.bandit]
exclude_dirs = ["tests", "alembic", ".venv", "venv"]
skips = ["B101", "B601"]  # assert_used, paramiko_calls

# ==================== Project Metadata ====================
[project]
name = "jobt-ai-career-coach"
version = "1.0.0"
description = "AI-powered interview practice platform"
requires-python = ">=3.11"
authors = [
    {name = "Micheal Olorundare", email = "contact@jobt.ai"}
]
maintainers = [
    {name = "Micheal Olorundare", email = "contact@jobt.ai"}
]
license = {text = "MIT"}
readme = "README.md"
keywords = ["interview", "ai", "career", "fastapi"]

[project.urls]
Homepage = "https://jobt.ai"
Documentation = "https://docs.jobt.ai"
Repository = "https://github.com/yourusername/jobt-ai"
"Bug Tracker" = "https://github.com/yourusername/jobt-ai/issues"
Changelog = "https://github.com/yourusername/jobt-ai/blob/main/CHANGELOG.md"